Rename directory to 'matrix' and update file paths
- Rename directory from 100k-rows-10-cols to matrix
- Update index.html title and script references for new structure
- Rename all related files (Toolbar.tsx, systemjs.config.js, etc.) to match new directory
- Update test specs to reference new file paths
- Adjust script URLs in HTML to reflect renamed directory structure
