+++ testing/performance/benchmarking.ts
@@ -162,7 +162,7 @@ const gridToChartsMap: Record<Version, Version> = {
     'v31.2.0': 'v9.2.0',
 } as const;
 
-const getCdnUrl = (pkg: string, version: Version, path: `/${string}` = `/dist/${pkg}.js`) => {
+export const getCdnUrl = (pkg: string, version: Version, path: `/${string}` = `/dist/${pkg}.js`) => {
     return `${knownUrlsProxy[version]}/files/${pkg}${path}`;
 };
